Remove proxy, separate app and api servers
