+++ 01-Login/server.js
@@ -9,8 +9,4 @@ const port = process.env.SERVER_PORT || 3000;
 app.use(morgan("dev"));
 app.use(express.static(join(__dirname, "build")));
 
-app.use((_, res) => {
-  res.sendFile(join(__dirname, "build", "index.html"));
-});
-
 app.listen(port, () => console.log(`Listening on port ${port}`));
