+++ 02-Calling-an-API/Dockerfile
@@ -32,9 +32,12 @@ RUN yarn install --production
 COPY --from=build /app/build ./build
 COPY --from=build /app/src/auth_config.json ./src/auth_config.json
 COPY --from=build /app/server.js .
+COPY --from=build /app/api-server.js .
 
 EXPOSE 3000
+EXPOSE 3001
 ENV SERVER_PORT=3000
+ENV API_PORT=3001
 ENV NODE_ENV production
 
-CMD ["node", "server.js"]
+CMD ["yarn", "prod"]
